ETL_project/requirements.txt
pandas
sqlalchemy
psycopg2-binary
geoip2
user-agents
